feat: add visual markers for position and accuracy
- Add Marker import from maplibre-gl
- Create position marker (blue dot with white border)
- Create accuracy circle (semi-transparent blue)
- Add show/hide marker methods
- Initialize markers on control creation
- Clean up markers on control removal
- Temporarily show markers on click for testing

The markers match the visual style of MapLibre's native GeolocateControl.
