--- src/index.complete.ts
@@ -1,7 +1,0 @@
-// Entry point for the library
-export { MockGeolocateControl } from "./MockGeolocateControl";
-export type {
-  MockGeolocateControlOptions,
-  GeolocateEventData,
-  OutOfMaxBoundsEventData,
-} from "./types";
