--- src/style.complete.css
@@ -1,188 +1,0 @@
-:root {
-    font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
-    line-height: 1.5;
-    font-weight: 400;
-
-    color-scheme: light dark;
-    color: rgba(255, 255, 255, 0.87);
-    background-color: #242424;
-
-    font-synthesis: none;
-    text-rendering: optimizeLegibility;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-}
-
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    margin: 0;
-    padding: 0;
-    width: 100vw;
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-}
-
-#map {
-    flex: 1;
-    width: 100%;
-}
-
-.header {
-    background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-    color: white;
-    padding: 1rem 1.5rem;
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    z-index: 10;
-}
-
-.header-content {
-    flex: 1;
-}
-
-.header-title {
-    font-size: 1.5rem;
-    font-weight: 700;
-    margin: 0 0 0.25rem 0;
-}
-
-.header-subtitle {
-    font-size: 0.875rem;
-    opacity: 0.95;
-    margin: 0;
-}
-
-.github-link {
-    display: inline-flex;
-    align-items: center;
-    gap: 0.5rem;
-    background: rgba(255, 255, 255, 0.2);
-    color: white;
-    text-decoration: none;
-    padding: 0.5rem 1rem;
-    border-radius: 0.375rem;
-    font-size: 0.875rem;
-    font-weight: 500;
-    transition: all 0.2s ease;
-    backdrop-filter: blur(10px);
-}
-
-.github-link:hover {
-    background: rgba(255, 255, 255, 0.3);
-    transform: translateY(-2px);
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-}
-
-.github-icon {
-    width: 20px;
-    height: 20px;
-}
-
-/* Demo Controls Panel */
-.demo-controls {
-    position: absolute;
-    top: 100px;
-    left: 10px;
-    background: white;
-    padding: 16px;
-    border-radius: 8px;
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
-    max-width: 280px;
-    font-size: 14px;
-    color: #333;
-    z-index: 1;
-}
-
-.demo-controls h3 {
-    margin: 0 0 12px 0;
-    font-size: 16px;
-    font-weight: 600;
-    color: #1a1a1a;
-}
-
-.demo-controls .section {
-    margin-bottom: 16px;
-    padding-bottom: 16px;
-    border-bottom: 1px solid #e0e0e0;
-}
-
-.demo-controls .section:last-child {
-    margin-bottom: 0;
-    padding-bottom: 0;
-    border-bottom: none;
-}
-
-.demo-controls label {
-    display: block;
-    margin-bottom: 4px;
-    font-weight: 500;
-    color: #555;
-}
-
-.demo-controls button {
-    display: block;
-    width: 100%;
-    padding: 8px 12px;
-    margin: 4px 0;
-    background: #2196f3;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    font-size: 13px;
-    transition: background 0.2s;
-}
-
-.demo-controls button:hover {
-    background: #1976d2;
-}
-
-.demo-controls button.secondary {
-    background: #f5f5f5;
-    color: #333;
-}
-
-.demo-controls button.secondary:hover {
-    background: #e0e0e0;
-}
-
-.demo-controls input[type="range"] {
-    width: 100%;
-    margin: 8px 0;
-}
-
-.demo-controls .value-display {
-    display: inline-block;
-    min-width: 50px;
-    color: #1976d2;
-    font-weight: 500;
-}
-
-.demo-controls .preset-buttons {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 8px;
-}
-
-.demo-controls .preset-buttons button {
-    font-size: 12px;
-    padding: 6px 8px;
-}
-
-.demo-controls .checkbox-wrapper {
-    display: flex;
-    align-items: center;
-    margin: 8px 0;
-}
-
-.demo-controls .checkbox-wrapper input {
-    margin-right: 8px;
-}
